+++ srcs/libc/f_string/src/f_print_variadic.c
@@ -34,7 +34,9 @@
 #include <f_string/f_print.h>
 #include <f_string/f_print_fd.h>
 
-static void	uf_print_treat_argument(va_list *ap, const char **fmt, size_t fd)
+#include <stdlib.h>
+
+static void	uf_print_treat_argument(va_list *ap, const char **fmt, int fd)
 {
 	*fmt = *fmt + 1;
 	if (**fmt == 'd' || **fmt == 'i')
@@ -53,7 +55,7 @@ static void	uf_print_treat_argument(va_list *ap, const char **fmt, size_t fd)
 		uf_print_color_fx(va_arg(*ap, int), 1);
 }
 
-void		uf_print_variadic_fd(size_t fd, const char *fmt, ...)
+void		uf_print_variadic_fd(int fd, const char *fmt, ...)
 {
 	va_list		ap;
 	const char	*tmp;
