+++ srcs/libc/string/src/s_string_print.c
@@ -45,7 +45,7 @@ void	f_string_print_memory(const t_string *v_this, const char *name)
 	uf_print_memory(v_this->v_str, v_this->v_capacity);
 }
 
-void	f_string_print_fd(const t_string *v_this, size_t fd)
+void	f_string_print_fd(const t_string *v_this, int fd)
 {
 	write(fd, v_this->v_str, v_this->v_size);
 }
